Clean up discord/frankerfacez/imgur resolvers
