--- internal/resolvers/frankerfacez/static_responses.go
@@ -1,14 +1,0 @@
-package frankerfacez
-
-import (
-	"net/http"
-
-	"github.com/Chatterino/api/pkg/resolver"
-)
-
-var (
-	emoteNotFoundResponse = &resolver.Response{
-		Status:  http.StatusNotFound,
-		Message: "No FrankerFaceZ emote with this id found",
-	}
-)
