+++ internal/resolvers/imgur/loader_test.go
