+++ internal/resolvers/imgur/resolver_test.go
